requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
jinja2==3.1.2
pydantic-settings==2.1.0
google-genai==0.2.1
python-dotenv==1.0.0
